+++ src/main/java/com/metalancer/backend/orders/entity/OrderPaymentEntity.java
@@ -10,6 +10,7 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.OneToOne;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.time.ZoneId;
 import java.util.Date;
@@ -48,7 +49,7 @@ public class OrderPaymentEntity extends BaseEntity {
     private String currency;
 
     @Column(nullable = false)
-    private Double paymentPrice;
+    private BigDecimal paymentPrice;
 
     @Column(nullable = false)
     private LocalDateTime purchasedAt;
@@ -60,7 +61,8 @@ public class OrderPaymentEntity extends BaseEntity {
     @Builder
     public OrderPaymentEntity(OrdersEntity ordersEntity, String impUid, String orderNo,
         String type, String title,
-        String method, Double paymentPrice, String currency, Date purchasedAt, String receiptUrl,
+        String method, BigDecimal paymentPrice, String currency, Date purchasedAt,
+        String receiptUrl,
         String paidStatus, String pgTid) {
         this.ordersEntity = ordersEntity;
         this.impUid = impUid;
