+++ src/main/java/com/metalancer/backend/orders/repository/ProductsSalesRepository.java
@@ -3,11 +3,12 @@ package com.metalancer.backend.orders.repository;
 import com.metalancer.backend.common.constants.CurrencyType;
 import com.metalancer.backend.orders.entity.ProductsSalesEntity;
 import com.metalancer.backend.users.entity.CreatorEntity;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public interface ProductsSalesRepository {
 
-    Double getTotalPriceByCreatorAndDate(CreatorEntity creatorEntity, LocalDateTime date,
+    BigDecimal getTotalPriceByCreatorAndDate(CreatorEntity creatorEntity, LocalDateTime date,
         LocalDateTime startOfNextDay, CurrencyType currency);
 
     int getSalesCntByCreatorAndDate(CreatorEntity creatorEntity, LocalDateTime date,
